requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.19.1
sox==1.4.0
